+++ src/lib/webcam.js
@@ -1,6 +1,6 @@
 //const enumerateDevices = require('enumerate-devices')
 
-module.exports = function (deviceId) {
+export default function (deviceId) {
   return navigator.mediaDevices.enumerateDevices()
     .then(devices => devices.filter(devices => devices.kind === 'videoinput'))
     .then(cameras => {
@@ -15,6 +15,9 @@ module.exports = function (deviceId) {
     })
     .then(stream => {
       const video = document.createElement('video')
+      video.setAttribute('autoplay', '')
+      video.setAttribute('muted', '')
+      video.setAttribute('playsinline', '')
       //  video.src = window.URL.createObjectURL(stream)
       video.srcObject = stream
       return new Promise((resolve, reject) => {